build: Exclude unnecessary fields when using --disable-v0
