+++ .github/workflows/tests.yml
@@ -16,13 +16,6 @@ jobs:
           sudo apt-get update -qq
           sudo apt-get install -qq lcov linux-libc-dev libuv1-dev btrfs-progs xfsprogs zfsutils-linux
 
-    - name: Amalgamation
-      run: |
-          git clone --depth 1 https://github.com/edlund/amalgamate.git
-          export PATH=$PATH:$PWD/amalgamate
-          amalgamate.py --config=amalgamation.json --source=$(pwd)
-          gcc raft.c -c -D_GNU_SOURCE -DHAVE_LINUX_AIO_ABI_H -Wall -Wextra -Wpedantic -fpic
-
     - name: Configure
       run: |
           autoreconf -i
@@ -35,6 +28,13 @@ jobs:
     - name: Build
       run: |
           make -j$(nproc --all)
+
+    - name: Amalgamation
+      run: |
+          git clone --depth 1 https://github.com/edlund/amalgamate.git
+          export PATH=$PATH:$PWD/amalgamate
+          amalgamate.py --config=amalgamation.json --source=$(pwd)
+          gcc raft.c -c -D_GNU_SOURCE -DHAVE_LINUX_AIO_ABI_H -Wall -Wextra -Wpedantic -fpic
 
     - name: Test
       run: |
