+++ src/restore.c
@@ -5,7 +5,6 @@
 #include "entry.h"
 #include "err.h"
 #include "log.h"
-#include "snapshot.h"
 #include "tracing.h"
 
 #define tracef(...) Tracef(r->tracer, __VA_ARGS__)
